[3P Libraries] Split Projector plugin into plugin + library.
This moves the vast majority of the Projector logic into a separate jar in the
libraries section; the plugin is just a thin wrapper around this jar.  All
files in the library jar are in the org.micromanager.projector.internal
package.

Additionally, the build system was updated to allow plugins to depend on
library jars.
