+++ buildscripts/Ant/unixprops.xml
@@ -20,4 +20,6 @@
 	<property name="mm.java.lib.mmcorej" location="${mm.basedir}/MMCoreJ_wrap/MMCoreJ.jar"/>
 	<property name="mm.java.lib.mmstudio" location="${mm.basedir}/mmstudio/MMJ_.jar"/>
 	<property name="mm.java.lib.acq-engine" location="${mm.basedir}/acqEngine/MMAcqEngine.jar"/>
+
+	<property name="mm.java.mmlibraries" location="${mm.basedir}/libraries"/>
 </project>
