+++ buildscripts/Ant/windowsprops.xml
@@ -95,6 +95,8 @@
 	<property name="mm.java.lib.mmstudio" location="${mm.java.outdir}/MMJ_.jar"/>
 	<property name="mm.java.lib.acq-engine" location="${mm.java.outdir}/MMAcqEngine.jar"/>
 
+	<property name="mm.java.mmlibraries" location="${mm.java.outdir}/libraries"/>
+
 	<!-- Other directories -->
 	<property name="mm.scripts.srcdir" location="${mm.basedir}/scripts"/>
 	<property name="mm.scripts.installdir" location="${mm.installdir}/scripts"/>
