+++ libraries/ExampleLibrary/README.txt
@@ -1,3 +1,5 @@
 This directory contains an extremely bare-bones example library. If you want to
 build this library, you will need to rename the file "rename-to-build.xml" to
-"build.xml" (without the quotation marks) first.
+"build.xml" (without the quotation marks) first, and add "ExampleLibrary.jar"
+to the nobase_mmlibrary_DATA section of the Makefile.am file (which is in the
+parent directory to this file).
