+++ libraries/Makefile.am
@@ -3,7 +3,7 @@
 # Note: You will fall into an infinite loop if the JAR filenames do not match
 # the JARs generated by Ant.
 nobase_mmlibrary_DATA = \
-	ExampleLibrary.jar
+	Projector.jar
 
 .PHONY: libraries.stamp
