+++ libraries/Projector/build.xml
@@ -1,0 +1,3 @@
+<project name="Projector" default="jar">
+	<import file="../javalibrarybuild.xml"/>
+</project>
