+++ libraries/Projector/src/main/java/org/micromanager/projector/internal/MosaicSequencingFrame.form
