+++ libraries/Projector/src/main/java/org/micromanager/projector/internal/OnStateListener.java
@@ -15,7 +15,7 @@
 //               INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES.
 
 
-package org.micromanager.projector;
+package org.micromanager.projector.internal;
 
 public interface OnStateListener {
    public void stateChanged(boolean onState);
