+++ libraries/Projector/src/main/java/org/micromanager/projector/internal/SLM.java
@@ -14,7 +14,7 @@
 //               CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
 //               INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES.
 
-package org.micromanager.projector;
+package org.micromanager.projector.internal;
 
 import ij.gui.PolygonRoi;
 import ij.gui.Roi;
