+++ plugins/Projector/build.xml
@@ -1,3 +1,3 @@
-<project name="Projector" default="jar">
+<project name="ProjectorPlugin" default="jar">
 	<import file="../javapluginbuild.xml"/>
 </project>
