--- plugins/Projector/src/main/java/org/micromanager/projector/Utils.java
@@ -1,28 +1,0 @@
-
-package org.micromanager.projector;
-
-import ij.process.FloatPolygon;
-import java.awt.Polygon;
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- *
- * @author nico
- */
-public class Utils {
-
-   public static List<Polygon> FloatToNormalPolygon(List<FloatPolygon> floatPolygons) {
-      // manually copy from FloatPolygon to Polygon
-      List<Polygon> roiPolygons = new ArrayList<Polygon>();
-      for (FloatPolygon fp : floatPolygons) {
-         Polygon p = new Polygon();
-         for (int i = 0; i < fp.npoints; i++) {
-            p.addPoint((int) (0.5 + fp.xpoints[i]),
-                    (int) (0.5 + fp.ypoints[i]));
-         }
-         roiPolygons.add(p);
-      }
-      return roiPolygons;
-   }
-}
